- Fixed pthread linkage error on gnuc - Fixed cmake configuration - Fixed asrun compilation on msvc6 - Worked on debugger
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@885 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/angelscript/source/as_context.cpp
@@ -3510,7 +3510,7 @@ bool asCContext::IsVarInScope(asUINT varIndex, asUINT stackLevel)
 	}
 
 	// First determine if the program position is after the variable declaration
-	if( (int)func->variables.GetLength() <= varIndex ) return false;
+	if( func->variables.GetLength() <= varIndex ) return false;
 	if( func->variables[varIndex]->declaredAtProgramPos > pos ) return false;
 
 	asUINT declaredAt = func->variables[varIndex]->declaredAtProgramPos;
